move slackcat struct and methods to own file, rename stream -> queue.go
